# Revert the glob_files_in_dir optimization