--- a/lib/rubygems/util.rb
+++ b/lib/rubygems/util.rb
@@ -99,11 +99,7 @@ module Gem::Util
   # returning absolute paths to the matching files.
 
   def self.glob_files_in_dir(glob, base_path)
-    if RUBY_VERSION >= "2.5"
-      Dir.glob(glob, base: base_path).map! {|f| File.expand_path(f, base_path) }
-    else
-      Dir.glob(File.expand_path(glob, base_path))
-    end
+    Dir.glob(File.expand_path(glob, base_path))
   end
 
   ##