@casperisfine 2020-07-06
Looking at the original PR (https://github.com/rubygems/rubygems/pull/971), it's possible that there's something missing in my benchmark.

I used theses calling values because that's what I saw in the boot profile, sounds like the original intent was to parse the entire load path for minitest plugins, the different pattern might change the performance profile. I'll try to see if I can reproduce the old 10X difference.