@deivid-rodriguez 2021-08-09
I adapted the benchmark to the original pattern, and it seems about 50% faster the way it's done now, so I'll close this.

```ruby
require "benchmark/ips"

class Hello
  def self.find_files_from_load_path_with_base_arg glob # :nodoc:
    glob_with_suffixes = "#{glob}#{Gem.suffix_pattern}"
    $LOAD_PATH.map { |load_path|
      Dir.glob(glob_with_suffixes, base: load_path).map! {|f| File.expand_path(f, load_path) }
    }
  end

  def self.find_files_from_load_path_without_base_arg glob # :nodoc:
    glob_with_suffixes = "#{glob}#{Gem.suffix_pattern}"
    $LOAD_PATH.map { |load_path|
      Dir.glob(File.expand_path(glob_with_suffixes, load_path))
    }
  end
end

search = "minitest/*_plugin.rb"

Benchmark.ips do |x|
  x.report('with base arg') do
    Hello.find_files_from_load_path_with_base_arg search
  end
  
  x.report('without base arg') do
    Hello.find_files_from_load_path_without_base_arg search
  end
  
  x.compare!
end
```

```
Warming up --------------------------------------
       with base arg     1.371k i/100ms
    without base arg   925.000  i/100ms
Calculating -------------------------------------
       with base arg     13.510k (± 1.7%) i/s -     68.550k in   5.075447s
    without base arg      9.298k (± 1.6%) i/s -     47.175k in   5.074872s

Comparison:
       with base arg:    13510.1 i/s
    without base arg:     9298.3 i/s - 1.45x  (± 0.00) slower
```